connected server with client, train and predict buttons now call bentoml service
